+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `Colourbars` test signal generator
 - Added `MovingBarOverlay` for test signal generators
 - Added `mediagrains.numpy` sublibrary for handling video grains as numpy arrays, in python 3.6+
+- Added `PSNR` option to grain compare.
 
 ## 2.5.3
 - BUGFIX: IOBytes doesn't quite fulfil bytes-like contracts, but can be converted to something that does